% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base2gisu.R
\name{base2gisu}
\alias{base2gisu}
\title{Title Convert the base information from P table mapped with risk rate table to Gisu}
\usage{
base2gisu(base3, r)
}
\arguments{
\item{base3}{table with required variables of seq, 나이, fo.##x, 보험가입금액, Q1000}

\item{r}{interest rate}
}
\value{
Gisu table with generated variables of t, ##x such as Dx, Nx, Cx, Mx
}
\description{
Title Convert the base information from P table mapped with risk rate table to Gisu
}
