% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisu2prem.R
\name{gisu2prem}
\alias{gisu2prem}
\title{Title Convert/calculate Premiums of Gisu-table based on informations called base}
\usage{
gisu2prem(기수표, exp)
}
\arguments{
\item{기수표}{Gisu-table}

\item{exp}{expense table}
}
\value{
Premiums
}
\description{
Title Convert/calculate Premiums of Gisu-table based on informations called base
}
