% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gisu2reserve.R
\name{gisu2reserve}
\alias{gisu2reserve}
\title{Title Convert/calculate Gisu-table to Reserve table}
\usage{
gisu2reserve(기수표, exp)
}
\arguments{
\item{기수표}{Gisu}

\item{exp}{}
}
\value{
Reserve table #
}
\description{
Title Convert/calculate Gisu-table to Reserve table
}
